+++ src/joi/object.ts
@@ -1,11 +1,9 @@
 
-
-
-import { createJoiItem, JoiObject, JoiSchema } from "./types";
-import { JSONSchema4 } from "json-schema";
+import { createJoiItem, JoiObject, JoiSchema } from './types';
+import { JSONSchema4 } from 'json-schema';
 import * as _ from 'lodash';
-import { resolveJSONSchema } from "./resolve";
-import { generateAnyJoi, generateJoi as generateJoi, JoiStatement, JoiSpecialChar, openJoi, closeJoi } from "./generate";
+import { resolveJSONSchema } from './resolve';
+import { generateAnyJoi, generateJoi as generateJoi, JoiStatement, JoiSpecialChar, openJoi, closeJoi } from './generate';
 
 function resolveProperties(schema: JSONSchema4): { [k: string]: JoiSchema } | undefined {
   const properties = schema.properties;
@@ -15,30 +13,46 @@ function resolveProperties(schema: JSONSchema4): { [k: string]: JoiSchema } | un
 
   return _.mapValues(properties, (property, key) => {
     const joiSchema = resolveJSONSchema(property);
-    joiSchema.required = (schema.required && schema.required.includes(key));
+    // https://json-schema.org/understanding-json-schema/reference/object.html#required-properties
+    const required = (schema.required && schema.required.includes(key));
+    if (required) {
+      joiSchema.required = true;
+    }
     return joiSchema;
   });
 }
 
 export function resolveJoiObjectSchema(schema: JSONSchema4): JoiObject {
   const joiSchema = createJoiItem('object') as JoiObject;
+  // https://json-schema.org/understanding-json-schema/reference/object.html#properties
   joiSchema.keys = resolveProperties(schema);
-  const additionalProperties = schema.additionalProperties
-  if (additionalProperties === true) {
-    joiSchema.unknown = true;
-  } else if (additionalProperties) {
+  let additionalProperties = schema.additionalProperties;
+  if (additionalProperties === undefined) {
+    additionalProperties = true;
+  }
+  if (typeof additionalProperties === 'boolean') {
+    joiSchema.unknown = additionalProperties;
+  } else {
     joiSchema.pattern = {
       pattern: '^',
       schema: resolveJSONSchema(additionalProperties),
     };
   }
+
+  // https://json-schema.org/understanding-json-schema/reference/object.html#size
   joiSchema.min = schema.minProperties;
   joiSchema.max = schema.maxProperties;
+
+  // TODO: Dependencies
+  //   https://json-schema.org/understanding-json-schema/reference/object.html#dependencies
+
+  // TODO: Pattern Properties
+  //   https://json-schema.org/understanding-json-schema/reference/object.html#pattern-properties
   return joiSchema;
 }
 
 export function generateObjectJoi(schema: JoiObject, level: number = 0): JoiStatement[] {
-  let content: JoiStatement[] = openJoi(['Joi.object()']);
+  const content: JoiStatement[] = openJoi(['Joi.object()']);
 
   const keys = schema.keys;
   if (keys) {
@@ -50,9 +64,9 @@ export function generateObjectJoi(schema: JoiObject, level: number = 0): JoiStat
     _.keys(keys).forEach((key) => {
       let printKey = key;
       if (key.includes(' ') || key.includes('-')) {
-        printKey = '\'' + printKey + '\''
+        printKey = '\'' + printKey + '\'';
       }
-      content.push(printKey)
+      content.push(printKey);
       content.push(JoiSpecialChar.COLON);
       content.push(...generateJoi(keys[key], level + 1));
       content.push(JoiSpecialChar.COMMA);
@@ -62,9 +76,7 @@ export function generateObjectJoi(schema: JoiObject, level: number = 0): JoiStat
       JoiSpecialChar.CLOSE_PAREN,
     ]);
   }
-  if (schema.unknown) {
-    content.push('.unknown()');
-  }
+
   if (schema.max !== undefined) {
     content.push(`.max(${schema.max})`);
   }
@@ -76,7 +88,7 @@ export function generateObjectJoi(schema: JoiObject, level: number = 0): JoiStat
     const pattern = schema.pattern.pattern;
     const subSchema = schema.pattern.schema;
     content.push(...[
-      '.pattern', 
+      '.pattern',
       JoiSpecialChar.OPEN_PAREN,
     ]);
     if (typeof pattern === 'string') {